Conflicts:
	configure
	src/cc1/Makefile
	src/cc1/Makefile.dep
	src/cc1/cc1.c
	src/cc1/out/out.c
	src/cc1/out/x86.c
	src/cpp2/main.c
	src/ucc/ucc.c
	src/util/platform.h

+++ test/initialisation/array/global_compile_time2.c
@@ -1,6 +1,6 @@
 int i;
 
-int x[] = {
+long x[] = {
 	&i
 };
